# Cache repeated parameter datatype queries for the same model