@patrickvonplaten 2023-09-12
I'm sadly not seeing a big speed-up anymore (after having merged: https://github.com/huggingface/diffusers/pull/4994). Could you maybe double check once quickly if you still see the same speed-up with the current version of main?