@isidentical 2023-09-12
Yeah, my own benchmarks also show no noticable speed up with this PR on (I think this was because of non-meta devices and my sequence of optimizations also reduced the need for this). Closing this PR as its currently very good as is!